Notificaciones basicas exitosas y avanze de la pantalla configuracion

+++ app/src/main/java/com/example/neolink_app/clases/configuracion/Confvalues.java
@@ -10,13 +10,13 @@ public class Confvalues {
     public double N_SLEEP_TIME;
     public double N_START_HOUR;
     public double PORT_RQ;
-    public double SLEEP_TIME;
+    public int SLEEP_TIME;
     public String WIFI_PSSWD_DEFAULT;
     public String WIFI_SSID_DEFAULT;
     public double aux;
 
     public void confvalues(){}
-    public void confvalues(double BAT_H, double BAT_L, double BEEP_EN, double GPS_RQ, double NO_WIFI, double N_END_HOUR, double N_SLEEP_TIME, double N_START_HOUR, double PORT_RQ, double SLEEP_TIME, String WIFI_PSSWD_DEFAULT, String WIFI_SSID_DEFAULT,double aux){
+    public void confvalues(double BAT_H, double BAT_L, double BEEP_EN, double GPS_RQ, double NO_WIFI, double N_END_HOUR, double N_SLEEP_TIME, double N_START_HOUR, double PORT_RQ, int SLEEP_TIME, String WIFI_PSSWD_DEFAULT, String WIFI_SSID_DEFAULT,double aux){
         this.BAT_H = BAT_H;
         this.BAT_L = BAT_L;
         this.BEEP_EN = BEEP_EN;
